feat(filterable-query-dropdown): rename query-search-dropdown to filterable-query-dropdown
