+++ src/inputs/dropdown/filterable-query-dropdown/PFilterableQueryDropdown.vue
@@ -1,6 +1,6 @@
 <template>
     <div v-click-outside="hideMenu"
-         class="p-query-search-dropdown"
+         class="p-filterable-query-dropdown"
     >
         <p-search ref="targetRef"
                   :class="{'no-menu': querySearchState.menu ? querySearchState.menu.length === 0 : false}"
@@ -99,12 +99,12 @@ import { useContextMenuFixedStyle, useProxyValue } from '@/hooks';
 import { useQuerySearch } from '@/hooks/query-search';
 import PContextMenu from '@/inputs/context-menu/PContextMenu.vue';
 import type { FilterableDropdownMenuItem } from '@/inputs/dropdown/filterable-dropdown/type';
-import type { QuerySearchDropdownProps } from '@/inputs/dropdown/query-search-dropdown/type';
+import type { FilterableQueryDropdownProps } from '@/inputs/dropdown/filterable-query-dropdown/type';
 import type { KeyMenuItem, ValueMenuItem, QueryItem } from '@/inputs/search/query-search/type';
 import PSearch from '@/inputs/search/search/PSearch.vue';
 
-export default defineComponent<QuerySearchDropdownProps>({
-    name: 'PQuerySearchDropdown',
+export default defineComponent<FilterableQueryDropdownProps>({
+    name: 'PFilterableQueryDropdown',
     components: {
         PContextMenu,
         PSearch,
@@ -242,7 +242,7 @@ export default defineComponent<QuerySearchDropdownProps>({
 </script>
 
 <style lang="postcss">
-.p-query-search-dropdown {
+.p-filterable-query-dropdown {
     @apply w-full relative;
     .p-search {
         .input-container {
