+++ src/inputs/dropdown/filterable-query-dropdown/type.ts
@@ -1,7 +1,7 @@
 import type { ContextMenuFixedStyleProps } from '@/hooks';
 import type { KeyItemSet, QueryItem, ValueHandlerMap } from '@/inputs/search/query-search/type';
 
-export interface QuerySearchDropdownProps extends ContextMenuFixedStyleProps {
+export interface FilterableQueryDropdownProps extends ContextMenuFixedStyleProps {
     value: string;
     placeholder: string;
     focused: boolean;
